+++ packages/twenty-server/src/engine/metadata-modules/utils/__tests__/validate-metadata-name-validity.spec.ts
@@ -1,26 +1,24 @@
-import {
-  validateMetadataNameOrThrow,
-  InvalidStringException,
-  NameTooLongException,
-} from 'src/engine/metadata-modules/utils/validate-metadata-name.utils';
+import { InvalidStringException } from 'src/engine/metadata-modules/utils/exceptions/invalid-string.exception';
+import { NameTooLongException } from 'src/engine/metadata-modules/utils/exceptions/name-too-long.exception';
+import { validateMetadataNameValidityOrThrow } from 'src/engine/metadata-modules/utils/validate-metadata-name-validity.utils';
 
-describe('validateMetadataNameOrThrow', () => {
+describe('validateMetadataNameValidityOrThrow', () => {
   it('does not throw if string is valid', () => {
     const input = 'testName';
 
-    expect(validateMetadataNameOrThrow(input)).not.toThrow;
+    expect(validateMetadataNameValidityOrThrow(input)).not.toThrow;
   });
   it('throws error if string has spaces', () => {
     const input = 'name with spaces';
 
-    expect(() => validateMetadataNameOrThrow(input)).toThrow(
+    expect(() => validateMetadataNameValidityOrThrow(input)).toThrow(
       InvalidStringException,
     );
   });
   it('throws error if string starts with capital letter', () => {
     const input = 'StringStartingWithCapitalLetter';
 
-    expect(() => validateMetadataNameOrThrow(input)).toThrow(
+    expect(() => validateMetadataNameValidityOrThrow(input)).toThrow(
       InvalidStringException,
     );
   });
@@ -28,7 +26,7 @@ describe('validateMetadataNameOrThrow', () => {
   it('throws error if string has non latin characters', () => {
     const input = 'בְרִבְרִ';
 
-    expect(() => validateMetadataNameOrThrow(input)).toThrow(
+    expect(() => validateMetadataNameValidityOrThrow(input)).toThrow(
       InvalidStringException,
     );
   });
@@ -36,7 +34,7 @@ describe('validateMetadataNameOrThrow', () => {
   it('throws error if starts with digits', () => {
     const input = '123string';
 
-    expect(() => validateMetadataNameOrThrow(input)).toThrow(
+    expect(() => validateMetadataNameValidityOrThrow(input)).toThrow(
       InvalidStringException,
     );
   });
@@ -44,14 +42,15 @@ describe('validateMetadataNameOrThrow', () => {
     const inputWith63Characters =
       'thisIsAstringWithSixtyThreeCharacters11111111111111111111111111';
 
-    expect(validateMetadataNameOrThrow(inputWith63Characters)).not.toThrow;
+    expect(validateMetadataNameValidityOrThrow(inputWith63Characters)).not
+      .toThrow;
   });
   it('throws error if string is above 63 characters', () => {
     const inputWith64Characters =
       'thisIsAstringWithSixtyFourCharacters1111111111111111111111111111';
 
-    expect(() => validateMetadataNameOrThrow(inputWith64Characters)).toThrow(
-      NameTooLongException,
-    );
+    expect(() =>
+      validateMetadataNameValidityOrThrow(inputWith64Characters),
+    ).toThrow(NameTooLongException);
   });
 });
