+++ packages/twenty-server/src/engine/metadata-modules/utils/exceptions/invalid-string.exception.ts
@@ -1,0 +1,7 @@
+export class InvalidStringException extends Error {
+  constructor(string: string) {
+    const message = `String "${string}" is not valid`;
+
+    super(message);
+  }
+}
