+++ packages/twenty-server/src/engine/metadata-modules/utils/exceptions/name-not-available.exception.ts
@@ -1,0 +1,7 @@
+export class NameNotAvailableException extends Error {
+  constructor(name: string) {
+    const message = `Name "${name}" is not available`;
+
+    super(message);
+  }
+}
