+++ packages/twenty-server/src/engine/metadata-modules/utils/exceptions/name-too-long.exception.ts
@@ -1,0 +1,7 @@
+export class NameTooLongException extends Error {
+  constructor(string: string) {
+    const message = `String "${string}" exceeds 63 characters limit`;
+
+    super(message);
+  }
+}
